Forgot to save scene (notes on previous commits in description)
Turns out that there was this really annoying bug that took me all day to solve. For collision, I was boxcasting from the centre of the object instead of the centre of the collider, so when I tried to resize and move the collider while sliding, collision would break. Now it's fixed.